+++ app/src/main/java/com/example/moneymanager/data/entity/Budget.kt
@@ -17,8 +17,8 @@ import com.example.moneymanager.data.entity.enums.Currency
     )]
 )
 data class Budget(
-    @PrimaryKey(autoGenerate = true) val id: Int = 0,
-    @ColumnInfo(name = "account_id") val accountId: Int,
+    @PrimaryKey(autoGenerate = true) val id: Long = 0,
+    @ColumnInfo(name = "account_id") val accountId: Long,
     @ColumnInfo(name = "type_color") val typeColor: String,
     val amount: Double,
     @ColumnInfo(name = "period_date_start") val periodDateStart: Long,
