+++ app/src/main/java/com/example/moneymanager/data/entity/TransferWallet.kt
@@ -6,11 +6,11 @@ import androidx.room.PrimaryKey
 
 @Entity(tableName = "transfer_wallet")
 data class TransferWallet(
-    @PrimaryKey(autoGenerate = true) val id: Int = 0,
+    @PrimaryKey(autoGenerate = true) val id: Long = 0,
     val amount: Double,
     val description: String,
-    @ColumnInfo(name = "from_wallet") val fromWallet: Int,
-    @ColumnInfo(name = "to_wallet")val toWallet: Int,
+    @ColumnInfo(name = "from_wallet") val fromWallet: Long,
+    @ColumnInfo(name = "to_wallet")val toWallet: Long,
     val fee: Double,
     @ColumnInfo(name = "link_img") val linkImg: String,
     @ColumnInfo(name = "transfer_date") val transferDate: Long
