+++ app/src/main/java/com/example/moneymanager/data/entity/Wallet.kt
@@ -5,9 +5,7 @@ import androidx.room.Entity
 import androidx.room.ForeignKey
 import androidx.room.PrimaryKey
 import androidx.room.TypeConverters
-import com.example.moneymanager.core.CurrencyTypeConverter
 import com.example.moneymanager.core.WalletTypeConverter
-import com.example.moneymanager.data.entity.enums.Currency
 import com.example.moneymanager.data.entity.enums.WalletType
 
 @Entity(
@@ -19,9 +17,9 @@ import com.example.moneymanager.data.entity.enums.WalletType
     )]
 )
 data class Wallet(
-    @PrimaryKey(autoGenerate = true) val id: Int = 0,
+    @PrimaryKey(autoGenerate = true) val id: Long = 0,
     @ColumnInfo(name = "amount") val amount: Double,
-    @ColumnInfo(name = "account_id") val accountId: Int,
+    @ColumnInfo(name = "account_id") val accountId: Long,
     @TypeConverters(WalletTypeConverter::class)
     @ColumnInfo(name = "type_wallet_name") val typeWallet: WalletType,
 )
